+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # 0.5.0
 
-Released on ?.
+Released on Friday, September 11 2026.
 
 - Updated to use the AngleSharp.Css gradient model (#10)
 - Improved rendering of overflow with ellipsis (#11)
+- Improved rendering with non-default `box-sizing` value
 - Added support for CSS grid track sizing (#7)
 - Added support for pseudo elements and `content` declaration (#8)
 - Added support for `position: sticky` (#9)